change enterprise card to client side
