+++ app/enterprise/metadata.ts
@@ -1,0 +1,12 @@
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'SigNoz Enterprise | Built for Scale',
+  openGraph: {
+    title: 'SigNoz Enterprise | Built for Scale',
+    description:
+      'With SigNoz Enterprise, you get unmatched scale and flexibility along with advanced security, compliance, and support.',
+  },
+  description:
+    'With SigNoz Enterprise, you get unmatched scale and flexibility along with advanced security, compliance, and support.',
+}
